rename ref_builder.schema -> .plan
