+++ ref_builder/legacy/convert.py
@@ -8,7 +8,7 @@ from ref_builder.logs import configure_logger
 from ref_builder.models import Molecule
 from ref_builder.ncbi.client import NCBIClient
 from ref_builder.repo import Repo
-from ref_builder.schema import OTUSchema
+from ref_builder.plan import OTUSchema
 from ref_builder.utils import DataType, IsolateName, IsolateNameType
