+++ tests/test_repo.py
@@ -11,7 +11,7 @@ from ref_builder.resources import (
     RepoOTU,
     RepoSequence,
 )
-from ref_builder.schema import OTUSchema, Segment
+from ref_builder.plan import OTUSchema, Segment
 from ref_builder.utils import Accession, DataType, IsolateName, IsolateNameType
